+++ src/main/java/org/synyx/urlaubsverwaltung/application/application/ApplicationInteractionServiceImpl.java
@@ -478,7 +478,7 @@ class ApplicationInteractionServiceImpl implements ApplicationInteractionService
     }
 
     @Override
-    public Optional<Application> get(Integer applicationId) {
+    public Optional<Application> get(Long applicationId) {
         return applicationService.getApplicationById(applicationId);
     }
