+++ src/main/java/org/synyx/urlaubsverwaltung/calendar/DepartmentCalendarRepository.java
@@ -10,12 +10,12 @@ import java.util.Optional;
 @Repository
 interface DepartmentCalendarRepository extends CrudRepository<DepartmentCalendar, Long> {
 
-    Optional<DepartmentCalendar> findByDepartmentIdAndPerson(Integer departmentId, Person person);
+    Optional<DepartmentCalendar> findByDepartmentIdAndPerson(Long departmentId, Person person);
 
     Optional<DepartmentCalendar> findBySecretAndPerson(String secret, Person person);
 
     @Modifying
-    void deleteByDepartmentIdAndPerson(Integer departmentId, Person person);
+    void deleteByDepartmentIdAndPerson(Long departmentId, Person person);
 
     @Modifying
     void deleteByPerson(Person person);
