+++ src/main/java/org/synyx/urlaubsverwaltung/person/PersonUpdatedEvent.java
@@ -5,13 +5,13 @@ import org.springframework.lang.Nullable;
 
 public class PersonUpdatedEvent extends ApplicationEvent {
 
-    private final Integer personId;
+    private final Long personId;
     private final String personNiceName;
     private final String username;
     private final String email;
     private final boolean active;
 
-    PersonUpdatedEvent(Object source, Integer personId, String personNiceName, String username, @Nullable String email, boolean active) {
+    PersonUpdatedEvent(Object source, Long personId, String personNiceName, String username, @Nullable String email, boolean active) {
         super(source);
         this.personId = personId;
         this.personNiceName = personNiceName;
@@ -20,7 +20,7 @@ public class PersonUpdatedEvent extends ApplicationEvent {
         this.active = active;
     }
 
-    public Integer getPersonId() {
+    public Long getPersonId() {
         return personId;
     }
