+++ src/main/java/org/synyx/urlaubsverwaltung/person/web/PersonPermissionsViewController.java
@@ -40,7 +40,7 @@ public class PersonPermissionsViewController {
 
     @PreAuthorize(IS_OFFICE)
     @GetMapping("/person/{personId}/permissions")
-    public String showPersonPermissionsAndNotifications(@PathVariable("personId") Integer personId, Model model) throws UnknownPersonException {
+    public String showPersonPermissionsAndNotifications(@PathVariable("personId") Long personId, Model model) throws UnknownPersonException {
 
         final Person person = personService.getPersonByID(personId).orElseThrow(() -> new UnknownPersonException(personId));
 
@@ -53,7 +53,7 @@ public class PersonPermissionsViewController {
 
     @PreAuthorize(IS_OFFICE)
     @PostMapping("/person/{personId}/permissions")
-    public String editPersonPermissionsAndNotifications(@PathVariable("personId") Integer personId,
+    public String editPersonPermissionsAndNotifications(@PathVariable("personId") Long personId,
                                                         @ModelAttribute("person") PersonPermissionsDto personPermissionsDto, Errors errors,
                                                         RedirectAttributes redirectAttributes) throws UnknownPersonException {
